Support closer zoom-in levels

+++ src/OpenLoco/include/OpenLoco/Graphics/RenderTarget.h
@@ -12,11 +12,11 @@ namespace OpenLoco::Gfx
     struct RenderTarget
     {
         uint8_t* bits;
-        int16_t x;
-        int16_t y;
-        int16_t width;
-        int16_t height;
-        int16_t pitch; // note: this is actually (pitch - width)
+        int32_t x;
+        int32_t y;
+        int32_t width;
+        int32_t height;
+        int32_t pitch; // note: this is actually (pitch - width)
 
         Ui::Rect getUiRect() const;
     };
